src/aiBackend/requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==2.0.1
google==3.0.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pip==23.3
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
setuptools==65.5.0
shapely==2.0.2
six==1.16.0
soupsieve==2.5
urllib3==2.0.7
vertexai==0.0.1
Werkzeug==2.2.2